+++ src/main/java/org/cactoos/list/SumOfReals.java
@@ -23,45 +23,49 @@
  */
 package org.cactoos.list;
 
+import java.util.Iterator;
 import org.cactoos.Scalar;
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.Matchers;
-import org.junit.Test;
 
 /**
- * Test case for {@link SumOfReal}.
+ * Real total of numbers.
+ *
+ * <p>There is no thread-safety guarantee.
  *
  * @author Vseslav Sekorin (vssekorin@gmail.com)
  * @version $Id$
  * @since 0.9
- * @checkstyle JavadocMethodCheck (500 lines)
- * @checkstyle MagicNumberCheck (500 lines)
  */
-public final class SumOfRealTest {
+public final class SumOfReals implements Scalar<Double> {
 
-    @Test
-    public void withVarargsCtor() throws Exception {
-        MatcherAssert.assertThat(
-            new SumOfReal(
-                () -> 1.2,
-                () -> 2.5,
-                () -> 3.3
-            ).value(),
-            Matchers.closeTo(7.0, 0.0)
-        );
+    /**
+     * The iterable.
+     */
+    private final Iterable<Scalar<Number>> src;
+
+    /**
+     * Ctor.
+     * @param src Numbers
+     */
+    @SafeVarargs
+    public SumOfReals(final Scalar<Number>... src) {
+        this(new ArrayAsIterable<>(src));
     }
 
-    @Test
-    public void withIterCtor() throws Exception {
-        MatcherAssert.assertThat(
-            new SumOfReal(
-                new ArrayAsIterable<Scalar<Number>>(
-                    () -> 7.1,
-                    () -> 8.1,
-                    () -> 10.1
-                )
-            ).value(),
-            Matchers.closeTo(25.0, 0.3)
-        );
+    /**
+     * Ctor.
+     * @param src The iterable
+     */
+    public SumOfReals(final Iterable<Scalar<Number>> src) {
+        this.src = src;
+    }
+
+    @Override
+    public Double value() throws Exception {
+        final Iterator<Scalar<Number>> numbers = this.src.iterator();
+        Double result =  0.;
+        while (numbers.hasNext()) {
+            result += numbers.next().value().doubleValue();
+        }
+        return result;
     }
 }
